Sample model and minor fix
A sample text classification model which I had mentioned earlier was not
pushed then (somehow I missed it). Also, fix for args in
data_downloader.
Path is the location where both the csv file and the folder where data
is to be downloaded are kept.
Thus, path and dirname are two arguments instead of one.