Don't yield versions that appear to be malformed
I ran across a bunch of malformed `data:` URLs in the CDX API last night, and I'm sure we might eventually find other kinds of content that look wrong and will never correspond to actual content that you can retrieve from the memento API. In `list_versions()`, we look for patterns of bad content and drop them from the yielded results.

Note this logic is in `list_versions()` and *not* in `search_cdx()`, the lower-level call. I didn't put them there because I thought it made sense to keep that focused on simply parsing and yielding lines and iterating pages of results. We could definitely make it a keyword arg, though.