Try to make CDXClient generators less error-prone
This... got a little crazy. Because our CDXClient methods return iterators, it's easy to screw up by writing some code like this:

    def get_some_versions():
        with CDXClient() as cdx:
            return cdx.list_versions('http://somewhere.com')

    for versions in get_some_versions():
        # do something

That will not throw an error, but it will fail to take advantage of the session (because the session gets closed before the iterator even starts iterating). I'm also not sure what ways this might silently *fail*. The session continues to *function* even though most of its internal state and its pool has been ravaged (and the pool does not get auto repaired; I read through the source and checked).

So I went down a crazy path of trying to make this defer the closing of the session until all the things using it are done. But that got complicated, as you can see in this commit. I'm not sure this is worthwhile. Another approach to take might be to subclass session and make it raise an exception if you call `send` after calling `close`, just to make it obvious that you are hitting this subtly confusing aspect of the API.