Catch non-mementos in long redirect chains
Also has better error handling for redirect loops, when the loop is multiple requests or doesn't start with the first request.

This finally pushes the `retry_request` implementation beyond the pale, in my opinion. *Definitely* needs to be reimplemented in a requests/urllib3-native way, which would really simplify pretty much everything here.