Handle mementos that are redirects better
If a memento is a redirect, its target will rarely actually have been captured at the same time. In the previous change, we only accepted those targets if they were actual mementos. Now we allow for the case where the target of an actual memento redirect can be a redirect to a nearby-in-time memento.