Configure retry/backoff per WaybackSession
This lets us use different settings more easily in different parts of the app. We could definitely do some sort of generic retryable adapter or something, but this seemed the most straightforward thing for now. I tried using the built-in urllib3 `Retry` class, except we can't customize how it behaves in the presence of various headers, which is kind of important for Mementos. See this issue for more: https://github.com/urllib3/urllib3/issues/1445