Make initial pass much more aggressive
After adding the second pass for failed mementos and experimenting a bit, I've finally realized the real problem is that many mementos are just *extraordinarily* slow to load -- so much so that parts of Wayback appear to timeout with gateway errors internally. HOWEVER, those loads also appear to eventually succeed in the background and get cached, such that a request a bit later *will* succeed.

So! That suggests that the right strategy is really to fail reasonably fast on the first pass, which lets lots more mementos get loaded while Wayback is laboriously reanimating an extra slow one, and trying the extra-slow one again. And, in fact, when we do that, we see almost no failures at the end of the run.