Remove redundant domains from CDX query list
Wayback CDX searches are actually by SURT key, not URL, so there's no difference between `www.whatever.com`, `www2.whatever.com`, and `whatever.com`. This makes sure that, if we have all of those in the domain list, we reduce it down to only `whatever.com` so we aren't asking for the same domain's worth of pages multiple times.