Try batching Wayback memento loads
The strategy of failing fast and retrying at the end works well for small sets, but hasn't been panning out so well in production runs for large sets of 30,000+ mementos. That's almost certainly because we are trying so many that they are falling out of Wayback's cache before we get back around to them. Instead, do both stages in sets of 2,000.