Use CurlAsyncHTTPClient instead of the default AsyncHTTPClient
According to the docs: https://www.tornadoweb.org/en/stable/httpclient.html
`CurlAsyncHTTPClient` is better than the default `AsyncHTTPClient` in many ways:

* `curl_httpclient` is more likely to be compatible with sites that are
not-quite-compliant with the HTTP spec, or sites that use little-exercised
features of HTTP.
* `curl_httpclient` is faster.

The required code modification to use it is minimal.

`pyurl` is an additional requirement to use it. I also needed to install
system pkg `libcurl4-openssl-dev` to do that in my server (Ubuntu 14.04
LTS).