Add (and handle) special error for `robots.txt`
It turns out that the CDX API obeys `robots.txt` (so it won't list content it has that violates a `robots.txt` policy), even though much of the rest of Wayback does not. This adds a special error class for CDX queries that hit `robots.txt` violations and handles it as a warning in our ETL script.