Query individual URLs if we have few in a domain
Some domains in our dataset are *really* big and we aren't actually interested in an appreciable proportion of them. In those cases, it doesn't make sense to query for the whole domain and then filter results on our side (this still makes sense when we're interested in a lot of URLs at the domain or there aren't many snapshots of a domain). We also have a handful of URLs at giant domains that almost never make sense to query as a whole (e.g. youtube.com).

To handle the above cases, there is now a list of `NEVER_QUERY_DOMAINS` (domains we should never query as a whole) and a `MAX_QUERY_URLS_PER_DOMAIN` threshold for how many URLs a domain has to have before we just query the whole domain instead of the individual URLs. To make all this work, `_get_db_page_url_info()` now returns a list of URLs to query CDX for, some of which may be domains (`http://{domain}/*`) and some of which may be exact URLs. That makes `wm db list-domains` not make much sense anymore, so I may have to come back and clean that up. This is all super messy at this point anyhow :(