Get even looser on memento redirects
Hopefully the comments added here are descriptive enough of the nuttiness of this situation. I think a significant number of our non-playbackable URLs are because of situations where the target URL has the same SURT URL but a very different actual URL -- and this usually happens when the closest-in-time capture was not the next-in-time capture (which is the one that would have resulted from the redirect). At the end of the day, though, we are probably only hurting ourselves by being this strict. It seems like there are lots of pages we are just failing to record and compare as long as we don't accept this situation and take what the archive gives us.

On the flip-side, we could do a CDX search for the actual next-in-time match at this point, but that seems a little nuts and, more-over probably very slow. Not sure it would be worthwhile. (Also, super complex!)