Handle SIGINT/SIGTERM by stopping gracefully
Since we moved to a threaded model, this is pretty important. This is a little ugly, and I think we are past the point where I should have made classes for our worker threads instead of attempting to encapsulate everything in a function. But, it works reasonably well: send SIGINT (ctrl + c) or SIGTERM once to trigger a graceful shutdown, which stops fetching new info from CDX, stops fetching any mementos in the queue, and stops listening to the status of DB imports (but finishes *sending* any in-flight imports to the DB). A repeat of the same signal stops the process immediately.

We could probably do some more work to make the hard stop a *little* less abrupt (trigger exceptions to interrupt the threads, stop importing at all, etc.), but this code is already fairly twisted and overcomplicated, and that didn't seem worthwhile.