Try not chunking the initial list of URLs
We previously chunked the list of wayback memento URLs into sets of 2000, and ran a first set of retries on those before moving onto the next batch, then after all the batches are done, doing a final retry across all remaining failures. I suspect that might no longer be helpful or necessary so this commit tries removing those batches.

In several test runs against NOAA URLs this afternoon, it *seems* to have a slightly elevated error rate (1%-5%) vs what we're currently seeing (0.5%-2%), but that might be a peculiarity of those particular times, or current conditions on Wayback's side. I suspect it might be something like that, so we'll give this a go for a couple days of nightly runs and see what we get from our logs.