Check if detected encoding is known in _extract_encoding
Use `codecs.lookup` to check if the detected encoding is known to
python. If not, fallback to default. Its beter if all encoding related
logic is inside `_extract_encoding` and not have some of it in `_decode_body`
as it was earlier.

Also, avoid trying to decode body twice, which may be a bit faster.