Experiment: iterate all CDX results immediately
Instead of pulling on the CDX search only when needed, start iterating through all the results immediately and just shove them onto a queue. We occasionally have pauses that I believe are due to large/slow CDX queries, and those pauses would be unlikely if we just kept the buffer filled. It also means we don't have to worry (as much) about missing items while we page through results because new results might affect the paging.

My main concern here is increased memory usage, but I doubt we are loading so much that this will really be that high impact.