Avoid RuntimeError handling HTTPError differently
Don’t use `raise_error=False` and instead pull the response object out of
the error.

If the response is actually coming from a web archive,
allow error codes. The `Memento-Datetime` header indicates
the response is an archived one, and not an actual failure
to respond with the desired content.