Remove special log handling for urllib3
We previously set special log levels for urllib3 because it was too noisy (it logged every retry as a warning, but Wayback's systems fail a lot and we *expect* having to retry often). However, we later stopped using urllib3's retry functionality because it couldn't distinguish between Wayback failures (which we want to retry) and mementos of failures (which are OK). So the custom logging setup is no longer necessary at all!

This also adds a few more comments to the retry code, and captures the above situation in the same place where we describe the problems with urllib3's built-in retry functionality that will hopefully get fixed in the future.