HOTFIX: fix truncated healtcheck script messages
Out healthcheck script was unintentionally stuck on a very old version of the Sentry SDK because it had to run on the same codebase as our Wayback imports, which were part of a very long-lived branch. When everything got merged, we wound up getting upgraded to a *much* newer version of the Sentry client package, and it turns out it now truncates strings at 512 characters. Of course, the healthcheck script outputs pretty long messages, so that's not OK.

The max length is hardcoded as a constant in `sentry.utils.MAX_STRING_LENGTH`, and there doesn't appear to be any setting or feature to override it. Instead, there is a comment in the code that says:

    # This is intentionally not just the default such that one can patch `MAX_STRING_LENGTH` and affect `strip_string`.

...so that's what I've gone ahead and done here. It doesn't feel great, so I also commented on a relevant post on Sentry's forums about it: https://forum.sentry.io/t/some-stack-traces-are-truncated/7309/4