Remove `Filter Test.ipynb` and related files
This notebook file doesn't do anything other than load and display a CSV, and we don't actually use that CSV for anything. It's not clear exactly how to make good use of it's contents, so it's probably worth cleaning out at this point to make the repo and its code more understandable.