Use 502 status code for hash mismatches
If the received content didn't match the hash a requester gave us, it's definitely not an unknown server error on our part -- either the requester had the wrong hash or the upstream response was bad. There's not HTTP status code that perfectly fits this, but at least calling it a 502 error (problem with an upstream response) is better than 500, and lets us be a bit smarter about error handling.

Fixes #491.