Handle expected errors properly in diffing server (#527)
Some time ago, someone started using `send_error(code, reason='some nice message')` to create errors in Tornado. I foolishly did not dive into the docs, and thought that was correct. It turns out it was not! The `reason` parameter is not your error message; it is the HTTP status text. That is, we were creating responses like:

    GET http://localhost:8888/html_token?a=https://google.commmmm&b=https://example.commmm

    HTTP/1.1 502 Could not fetch https://google.commmmm: [Errno 8] nodename nor servname provided, or not known
    Content-Type: application/json; charset=UTF-8
    Date: Thu, 05 Dec 2019 18:31:01 GMT

    {"code": 502, "error": "Could not fetch https://google.commmmm: [Errno 8] nodename nor servname provided, or not known"}

Check out that `HTTP/1.1 502 Oh my god this whole error message` line (it should be HTTP/1.1 502 Bad Gateway). *(Screams)*

It turns out `send_error()` has no way to send nicer messaging, and we were deeply abusing it. Tornado's built-in tool for this kind of thing is raising `HTTPError`, but it also has no provision for public error messaging, let alone the extra details we were adding. So I've created `PublicError` for this situation. The message it logs is publicly visible (you can also add `log_message` for internal/Sentry logging if you need a different message). It also allows for extra structured data.

Now you get:

    GET http://localhost:8888/html_token?a=https://google.commmmm&b=https://example.commmm

    HTTP/1.1 502 Bad Gateway
    Content-Type: application/json; charset=UTF-8
    Date: Thu, 05 Dec 2019 19:48:14 GMT

    {
        "code": 502,
        "error": "Could not fetch \"https://google.commmmm\": [Errno 8] nodename nor servname provided, or not known",
        "url": "https://google.commmmm",
        "cause": "[Errno 8] nodename nor servname provided, or not known"
    }

Much nicer. As a bonus, we are now raising errors, which results in nicer control flow, instead of sending the error and then having to remember to return.

Finally, this uses the private `log_message` feature from `PublicError/HTTPError` to make sure all our upstream errors get sent to Sentry with the same message (you can still get the differentiating details from the "events" view in Sentry), so they can actually get counted and graphed together ("how often are we failing to load the URLs that need diffing" is a tough question right now).