Track import errors by import ID (#553)
When the import script breaks things up into multiple individual imports, it's hard to figure out which imports triggered which error messages (if there were errors), and therefore hard to remediate those errors later. Instead of returning a list of all the errors from all the imports that we are monitoring, this now returns a dict mapping import IDs to errors, and displays it with one import job's error list per line.