Use a shared HTTPAdapter across all threads (#551)
Wayback has newly imposed rate limits on the memento API that we are running afoul of. This limits actual requests (rather than mementos) to an acceptable rate by sharing an HTTPAdapter and thus a single pool of connections across all threads. (NOTE: this rate is actually problematically slow for us and we are working with Wayback folks to raise it, but 10/second is what it is now.)

You can set a custom rate limit using the `WAYBACK_RATE_LIMIT`, which is the number of requests allowed per second.

This also rewrites the `rate_limited` context manager as a lock-like object for more portability/re-use. This also means we don't keep a global registry of rate limiters, which is probably good.