Stop diff server gracefully on SIGINT/SIGTERM (#577)
When sent a `SIGTERM` signal (how Kubernetes stops pods), the diff server's main process exits immediately, but doesn't clean up any of the child processes that are running diffs, leading to problems (and *very* slow shutdowns in Kubernetes as it waits and waits and waits for all the processes in the container to exit before giving up and sending `SIGKILL`). This adds graceful shutdown handling for both `SIGINT` and `SIGTERM`:

1. Stop handling new incoming connections
2. Stop running new diffs (or cease any in-process diffs if a signal is sent a second time)
3. Wait for responses to finish writing

Fixes #576.