Add preliminary support for Tornado's cURL client
The cURL HTTP client in Tornado is both faster and more reliable for complex HTTP scenarios. I'm hoping this might improve some of the weird errors we run into in production every so often.

This is just the basics -- we need to fix up error handling, URL validation, and add docs for installation (now we have a dependency on libcurl & OpenSSL, and you have to make sure you've built them the right way). We probably also need to make changes to (or at least test) the dockerfile.