Shut down the process pool safely
We've been seeing `OSError` getting raised from inside the differ process pool for a while, but it's always been really unclear what's actually going on. (It seemed to occur in conjunction with SSL errors, so for a while I had thought some really strange voodoo was going on.) It turns out I had naively trusted the documentation for `executor.shutdown()`, which is either unclear or incorrect. It implies `executor.shutdown(wait=False)` returns immediately, but doesn't tear down the executor until after all pending tasks are complete, just like `executor.shutdown(wait=True)`. This is not true! Setting `wait=False` will cause internal polling on the pending tasks to raise the `OSError`.

To work around this, I've added a thin shim that schedules the shutdown and waits on a thread (using the default executor). The result is awaitable, but not a coroutine, so you can trigger it and move on *or* `await` it if you need to know when it's done.