Remove tzlocal dependency (#598)
When updating tzlocal a couple weeks ago, I noticed we only use it in one place, and we are using it for a feature already supported by another dependency, python-dateutil. In addition, it turns out some authoritative voices recommend moving away from pytz, which tzlocal is designed to work on top of (https://github.com/pganssle/pytz-deprecation-shim).