Parse titles correctly from PDFs when importing
We used to try and extract the title of any and every document as if it were HTML. However, we import lots of things that *aren't* HTML (e.g. PDF, CSV, XLS) and trying to parse those as HTML can slow, pointless, or memory intensive (the importer recently crashed by running out of memory trying to parse a PDF this way). This change takes more care to only extract the titles from HTML and from PDFs, and adds proper PDF parsing so we actually *can* extract the titles instead of failing and wasting huge amounts of memory every time.