Set an import's media_type instead of metadata
We used to use `source_metadata` to set the media type and encoding information about a version, but web-monitoring-db has long since added top-level support for media type information. This updates the import code to use that, and also makes sure not to set `media_type` to an empty string, which is invalid (we used to do this, and imports would be rejected in some rare cases as a result).

Since we are processing a memento's `Content-Type` header in a more nuanced way now, this also means we might start storing extra media type parameters that we were leaving out before (we previously only captured the `charset` parameter).