Use a single DB client throughout import script
This allows VCR to properly track all requests to web-monitoring-db (turns out it was missing some before, I think because one of the clients was created on a separate thread). I've also simplified the test here to include slightly less data so it runs quickly (but still covers several real-world cases).