Fix parameter for `skip_unchanged_versions, speed up import test (#623)
It turns out we had a mixup that was disabling the `skip_unchanged_versions` parameter from being sent along to the database with an import!

This also makes sure we only use a single DB client throughout the import script. That allows VCR to properly track all requests to web-monitoring-db (turns out it was missing some before, I think because one of the clients was created on a separate thread). I've also simplified the test here to include slightly less data so it runs quickly (but still covers several real-world cases).