Keep support for simple AND curl HTTP clients
The cURL integration handles all kinds of things (and, most obviously, errors) pretty differently, so it seems like a good idea to make falling back to the simple client pretty easy (or until we've had cURL running in production for a while, at least). Use the `USE_SIMPLE_HTTP_CLIENT` environment variable to switch back to the simple client.