Upgrade Wayback to v0.3.0a1 (#643)
The first alpha version of Wayback v0.3.0 is out, and as its primary users (and biggest scale users AFAIK), we should get this running so we find out about any issues quickly. It would be good to have this running for our imports before really certifying a final version of Wayback v0.3 anyway.