Fix unplaybackable cache and retries (#652)
I recently noticed that the unplaybackable cache didn't seem to be used, and I'm not sure when it broke. While doing that, I also noticed that the secondary retry functionality was also broken (!) so I fixed that too.

Finally, I went ahead and added better logging for the case where we didn't actually import anything. Fixes #582.