Don't fail on malformed/encrypted/unreadable PDFs
It turns out the PyPDF2 code we added in #622 was not especially safe -- there are a variety of ways PDF parsing could fail (the file could use a newer format PyPDF2 does not understand, or it could simply be malformed). It also turns out quite a few PDFs out there are encrypted with an empty password (weird, but common enough that most PDF readers simply try it automatically for it if the file is encrypted). Finally, PyPDF2 only supports 2 of the standard PDF encryption algorithms, so decryption could also fail on a perfectly legitimate file.

This attempts to handle all those cases gracefully.