Handle malformed PDFs more comprehensively
This is the third round of improved error handling for PDFs. We were hitting errors with a few types of malformed PDFs, so I wound up just backing up and wrapping the whole function body here in a try...except block. I've also added a bunch of test cases. This *should* do a much better job.