Add retryability to DB client (#666)
Normally this might be something to wrap around the client, but because we'll soon have built-in iteration (#660), we need retries built-in so they happen at a lower level. When you create a client, you can now specify a number of retries or a more complex urllib3 `Retry` object.

To do this, I had to create a custom `Session` subclass. Now that we have that, it also made sense to move the timeout functionality there, too.

Fixes #659.