Log less and import better (#668)
Our import logs are *huge*, mainly because we log every memento that could not be played back. Since we've known for a while that a significant number of these non-playbackables are that way by design, it doesn't really make sense to log them all anymore. (If there are a lot of mementos of the same URL in a short timeframe, Wayback makes some of them inaccessible in order to increase cache efficiency, since most of them are likely to be the same.)

This also does some minor cleanup of vestigial arguments and so on in the script.