Add `--precheck` option to import script (#667)
This adds a new `--precheck` option to the import script, which causes it to load a list of all the versions we have in the database for the given timeframe before trying to import anything from Wayback. The idea here is to reduce duplicate loads and imports: we currently run the import daily, but cover a timeframe of several days so we don’t have huge data loss every time Wayback’s CDX indexing falls behind by a few days (which is frequent — it’s happening as I write this, in fact). That helps things when Wayback is having issues, but when all is well, it results in us loading *a lot* more data from Wayback than is really necessary, and creating imports that frequently consist of data we ultimately ignore because it’s already in the database.

It’s an option because it can be pretty slow or take lots of memory if the import covers a large timeframe. It also can’t be narrowed by URL pattern very easily right now ([there’s no index on the `capture_url` of versions](https://github.com/edgi-govdata-archiving/web-monitoring-db/blob/46561ae6eb52b0d923f7832100d161fc98667d0c/db/schema.rb#L153-L159) and querying pages then versions is complex, even slower, and doesn’t really help our main use case, which is the nightly import for *all* URLs).