Detect encoding correctly before extracting titles (#678)
While fixing some issues around prioritization in task sheets today, I noticed that we are not always correctly decoding content correctly before getting the page title, so we have some bad records. (For example, this page: https://www.cdc.gov/coronavirus/2019-ncov/travelers/after-travel-precautions-japanese.html)

This is mainly a problem caused by requests' frustrating encoding detection behavior (it won't tell you if if failed to get an encoding from headers and just sets a default that is often incorrect). This pulls in, but slightly modifies, the encoding detection logic from the server. We should also see about changing the behavior for the `encoding` attribute in the Wayback package.