Show every CDX record in import progress meter
One long-standing issue with the import script is that, because of the overly-complex retry queue logic, it only shows progress in terms of new versions loaded, and does not include skipped versions and errors in the count, leading to a lack of clarity in what's going on. The real question the progress bar should be answering is "how many CDX records have been processed?" Now that we've stopped using the retry queue, this is a start towards doing that.

The code here is *super* messy. There's a lot of crazy logic to unwind, and this script can be futzy to work with, so cleaning up the logic may take a little while. But this first commit *does* work, even if it's pretty ugly.