Show every CDX record in import progress meter (#680)
One long-standing issue with the import script is that, because of the overly-complex retry queue logic, it only shows progress in terms of new versions loaded, and does not include skipped versions and errors in the count, leading to a lack of clarity in what's going on. The real question the progress bar should be answering is "how many CDX records have been processed?" Now that we've stopped using the retry queue, this is a start towards doing that.

This is not very pretty, but helps answer monitoring-related questions more clearly (that's important because we are having some *fun* Wayback issues lately). There will be more to clean up about this over time.