Bump wayback to v0.3.0
There's a final, non-beta release of Wayback v0.3.0! Upgrade to that, and get rid of the timeout hacks.

Also reduce the `get_memento()` timeout to 45 seconds. The default 60 second timeout works and gives us a low error rate (0.15%-1.0%), but leads to import runs as long as 18 hours. Let's see whether reducing the timeout a bit shortens the runs without dramatically increasing errors.