Use explicit python version for Docker and Circle (#779)
We cache a virtualenv on CircleCI, and when there are new patch releases of Python and Circle upgrades, we have problems, because the cache's symlink for Python starts pointing to nowhere. This will help prevent that problem.

I figured it makes sense to lock down a specific version of Python in the Dockerfile while we’re at it.