Sniff media types when content-type is meaningless
When importing, we occasionally hit meaningless `Content-Type` headers like `application/octet-stream` or `application/x-download`, which servers often set in order to force a browser to download a file instead of display it. For these responses, we now sniff the media type according to a standard W3C/WHATWG approach. Note that the original `Content-Type` header is still stored in the `headers` field of the imported version, and we are only changing what winds up in the `media_type` field.