Add debug logs and possible threading fix
The production server this runs on sometimes gets stuck with things hanging, never getting to the stage where it reports results. I'm not sure quite what's going on there, but this adds some logging to help us see when different threads finish and moves some filtering into a thread with queues in case it might have been the source of the problem.