Correct configured DB client URLs and backoff (#843)
When configuring the DB client, you might specify a URL with a trailing slash (e.g. `https://abc.org/`) and wind up with it requesting URLs like `https://abc.org//api/v0/pages` (note the double slash). Our API handles this fine (yay), but it would be better if that didn't happen in the first place, so this fixes it.

While doing that, I also noticed we could do with some backoff when monitoring import statuses, and added that, too.