Don't use TQDM on non-TTYs
Normally, TQDM does not do anything if the output is not an interactive terminal (TTY), since the way it prints doesn't work very well in that context. However, we had some funky customizations that made it output in those environments anyway, because we were so often testing weird behavior over very long timespans, and having it output some kind of "heartbeat" was helpful. We haven't done that in a long time, and it's always made non-interactive logs messy. But now that those logs are going through Kubernetes to CloudWatch, it makes the logs basically unusable. This hack has long outlived its purpose.