Clean up logs for Kube/CloudWatch (#850)
Normally, TQDM does not do anything if the output is not an interactive terminal (TTY), since the way it prints doesn't work very well in that context. However, we had some funky customizations that made it output in those environments anyway (albeit at a lower rate), because we were so often investigating weird behavior over very long timespans in production, and having it output some kind of "heartbeat" was helpful. We haven't done that in a long time, and it's always made non-interactive logs messy. But now that those logs are going through Kubernetes to CloudWatch, it makes the logs basically unusable. This hack has long outlived its purpose.

I’ve also added logging for a final job timing/duration for convenience in archived logs.