Don't show progress bars if no TTY, take 2 (#851)
This is a follow-on to #850, which I botched. I thought I remembered TQDM not showing progress bars in non-TTY environments by default, but it turns out that is not true (or it's changed). Instead, you have to explicitly tell it to choose based on the environment with `disable=None` (instead of giving it a boolean value). This should actually remove the progress bars in production. (Also, this time I tested with Docker locally.)