Add v1 schema support to annotations import script (#853)
This project is technically deprecated, but I’m doing some work here to support final shutdown and archival of data (https://github.com/edgi-govdata-archiving/web-monitoring/issues/170).

The main goal here is to import old annotations (different schema than the import script was built to support) and some much newer ones that were never brought into the DB proper. I want them in the DB so we can export a nice SQLite archive that's easy for people to dig through, as opposed to collating data from a variety of sources.