Drop Python 3.9 support
The support matrix is too complicated here. We need some pieces that depend on boto, which depends on other things that require different versions of urllib3 on Python 3.9 than on Python 3.10. But VCR can only be compatible with urllib3 v1 OR v2, not both (it generates/reads different cassette formats). To keep things simple, I'm dropping Python 3.9 support. We mainly worried about old Pythons for Google Colab, and it is now on Python 3.11