Add `--archive-s3 <bucket>` option
This is pretty ugly, but it's a working first cut at the problem and fixes #663. This is suddenly more important since IA seems to be under a lot of load with peopole checking up on Trump-related website changes, and I'm seeing a lot more import-related failures on things that work fine on the processing side and fail to re-download in the DB importer. This will short-circuit that issue.