Add `--archive-s3 <bucket>` option (#855)
Add the ability to pre-upload content to our content-addressed storage bucket in S3 by setting the bucket name with the `--archive-s3 <bucket>` CLI option.

Normally, the DB will grab the content from whatever URL is specified when handling imports. But when we’re importing from S3, that means we pull it out of the Internet Archive twice: once to pre-process it here, and once after we send the import command to web-monitoring-db. That doubles the load on IA's systems, which is not great. With this change, we put it in our archival storage here, while pre-processing, and web-monitoring-db will see that it’s already there and not contact IA at all.